adjusting routes so we can subscribe to official zeta status page
