+++ package.json
@@ -24,6 +24,7 @@
     "typescript": "5.3.3"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.6"
+    "@types/bcryptjs": "^2.4.6",
+    "eslint": "^9.15.0"
   }
 }
